Prototype implementation of websocket-backed cross-context messaging
This allows creating message channels that can cross multiple browsing
contexts or script execution environments.

A simple channel pair can be created using the channel function e.g.

let [recvChan, sendChan] = await channel();

This generates a channel pair such that messages sent on sendChan will
arrive on recvChan.

Channels are multiple producer, single consumer, so there can only be
a single ReadChannel for each network, but multiple SendChannels.

On top of this underlying abstraction there's a higher level API which
provides known message structures. There is currently a command
message type where each command has an id, a method, a set of params,
and a respChannel which is a channel to use to reply to the command.

Currently two commands are provided: executeScript, which runs a
function in a remote context, and postMessage, which sends a message
to a remote context. Messages and script parameters are serialized
using a WebDriver BiDi like remote value serialization algorithm which
allows passing complex objects over the wire.

To work with multiple contexts, one takes the following actions:

* In the context that will recieve messages, use:

  let revChannel = await ctx_channel();

  This starts listening for messages on a channel defined by the
  testdriver context id (typically in the URL; maybe this should just
  be always in the URL which could also remove the testdriver.js
  dependency).

  If only executeScript is being used nothing more needs to be
  done. For postMessage, one can register message handlers using
  `recvChannel.addMessageHandler(fn)`.

* In the context that will send messages, use the uuid of the remote
  context to create a TestDriverRemote object. This will allow sending
  messages to the remote.

  let remote = TestDriverRemote(uuid);
  let result = remote.executeScript(selector =>
    document.querySelector(selector).textContent,
    ["body"])

  remote.postMessage("done");

Message channels are backed with websockets, which use the stash as
the data storage layer. It would also be possible to create a
HTTP-based backend, with POST to create a message and GET to receive a
message. For sending this would be a drop-in replacement but for
receiving it would require polling.