[selectmenu] Add basic keyboard support
Apply "controller" code to <selectmenu> parts for the following
behaviors:

- Enter key opens and closes the listbox.
- Spacebar opens the listbox (but does not close it -- this is
suspicious but consistent with Chromium <select>. Firefox's Enter and
Spacebar behavior for <select> is a bit different; we may want to
evaluate this in OpenUI).
- Up/Down Arrow move up/down through the list of options.

Event propagation is suppressed to prevent the default browser
behavior of scrolling the window on arrow keys or spacebar.

Bug: 1121840
Change-Id: I94c994ed8b381ef075683df807e852e9c92427ce