Add shared locks for writable streams
This CL implements the missing methods in WriteLockManager and hooks
them up to the file writer lifecycle. This enables shared write locks
when creating writable streams.

When a shared lock is taken by creating a writable stream, the creation
of access handles on the same file is prevented. Creating new writables
is permitted. Closing all writables releases the lock on the file.

Bug: 1218431
Change-Id: I9c21b901ed0f5f331682c8a6b75db5e28674e5a3