[mathml] Embellished operators and stretch properties
This is part patch of a series of CL to implement MathML stretchy
operators as well as embellished operators [1]. It handles the cases
"If there is a block stretch size constraint or an inline stretch size
constraint then the element being laid out is an embellished operator"
from [2] and [3].

A typical use case is the following, the inner <mover> element is an
embellished operator (a over brace with the " k times" as over script).
Its core operator should stretch horizontally to cover the inline size
of the <mrow> element "x + x + … + x". Without that CL, the horizontal
over brace would just cover the size of the " k times" over script.

    <math>
      <mover>
        <mrow>
          <mi>x</mi>
          <mo>+</mo>
          <mi>x</mi>
          <mo>+</mo>
          <mo>…</mo>
          <mo>+</mo>
          <mi>x</mi>
        </mrow>
        <mover>
          <mo>⏞</mo>
          <mrow>
            <mi>k</mi><mtext>&nbsp;times</mtext>
          </mrow>
        </mover>
      </mover>
    </math>

The outer <mover> passes inline stretch constraint to its over script
per [3]. Now using the rule from [3] implemented in this CL, the inner
<mover> will just use that inline stretch constraint to stretch the
over brace.

Note, that elements can receive a stretch constraint orthogonal to the
direction they normally stretch their children. In the example below,
the <munder> is an embellished operator (a left parenthesis with a "1"
under script). Although it normally stretches its children horizontally,
the parenthesis is stretched vertically using the stretch constraint
received from its parent <mrow>.

  <math>
    <mrow>
      <munder>
        <mo>(</mo>
        <mn>1</mn>
      </munder>
      <munderover>
        <mn>2</mn>
        <mn>3</mn>
        <mn>4</mn>
      </munderover>
    </mrow>
  </math>

New tests integrated into this CL cover all possible combinations:
- third_party/stretch-along-block-axis-001.html is extended to test
  case when the <mo> is replaced with an embellished operator with
  <mrow> or <munderover>. It tests when the mrow and underover
  algorithms forward any inherited block stretch size constraint to
  their children. It also tests when the mrow algorithm calculates
  and passes a block stretch size constraint to embellished operator
  children.
- presentation-markup/scripts/underover-stretchy-00[23].html are copies
  of underover-stretchy-001 with the <mo> embellished with <mrow> and
  <munderover> respectively. They respectively test when the mrow and
  underover algorithms forward any inherited inline stretch size
  constraint to their children. They also test when the munderover
  algorithm calculates and passes an inline stretch size constraint to
  embellished operator children.

[1] https://w3c.github.io/mathml-core/#embellished-operators
[2] https://w3c.github.io/mathml-core/#algorithm-for-stretching-operators-along-the-block-axis
[3] https://w3c.github.io/mathml-core/#algorithm-for-stretching-operators-along-the-inline-axis

Bug: 6606, 1124298, 1124301

Change-Id: I447433924cf1495a4be23d81f406266f84c7141e