Fix border snapping/rounding for background-clip: padding-box
LayoutNG was rounding border-widths to the nearest integer rather than
towards zero (or negative infinity - the difference is not relevant in
this case).
Drop all "pixel snapping" code from NGPhysicalBoxFragment and clean away
code is rendered unused by that (like NGPixelSnappedPhysicalBoxStrut).
Add BoxPainterBase::ComputeSnappedBorders() and use that in
BoxPainterBase.
This is probably still not entirely consistent with how background
geometry is computed, but is hopefully a step in the right direction.

Bug: 1240617
Change-Id: Ia0724e0ee1c87386f2699c44ac117e237f3cad8a